♻️ Refactor component files
