+++ src/app/account/page.tsx
@@ -1,6 +1,6 @@
-import PageTitle from '@/components/atom/PageTitle'
-import AccountButtons from '@/components/organism/AccountButtons'
-import AccountInfo from '@/components/organism/AccountInfo'
+import PageTitle from '@/components/common/PageTitle'
+import AccountButtons from '@/components/user/buttons/AccountButtons'
+import AccountInfo from '@/components/user/AccountInfo'
 
 function page() {
   return (
