+++ src/app/book/[id]/page.tsx
@@ -1,8 +1,8 @@
 'use client'
 import { getBookInfo } from '@/apis/book'
-import PageTitle from '@/components/atom/PageTitle'
-import BookInfo from '@/components/molecule/BookInfo'
-import ReadingInfo from '@/components/organism/ReadingInfo'
+import PageTitle from '@/components/common/PageTitle'
+import BookInfo from '@/components/book/BookInfo'
+import ReadingInfo from '@/components/book/ReadingInfo'
 import { Button, RibbonBadge } from '@yeonsubaek/yeonsui'
 import { useRouter } from 'next/navigation'
 import { useEffect, useState } from 'react'
