+++ src/components/book/BookInfo.tsx
