+++ src/components/book/Library.tsx
@@ -1,7 +1,7 @@
 'use client'
 import { Segmented } from '@yeonsubaek/yeonsui'
-import Bookshelf from './Bookshelf'
-import BookList from '../molecule/BookCardList'
+import Bookshelf from './Bookshelf/Bookshelf'
+import BookList from './BookCard/BookCardList'
 import { useTranslation } from 'react-i18next'
 import { useEffect, useState } from 'react'
 import { getUserInfoApi } from '@/apis/user'
